Update(base-handler.js): game over 시 baseHp를 함께 보내도록 변경
Update(game-over.handler.js): 최종 baseHp를 화면에 출력하도록 변경
Refactor: displayModalMessage 함수 사용
Refactor: Audio 인스턴스들을 함수에서 분리

Update: S2C_GameOverNotification protobuf 메세지에 baseHp 필드 추가

Update: showMatchButton 함수 추가
Update: client version 상수 추가

Refactor: 불필요한 comment, console 제거
