+++ public/src/constants/constants.js
@@ -1,1 +1,2 @@
 export const SERVER_URL = 'http://13.209.73.219:5555';
+export const CLIENT_VERSION = '1.0.0';
