+++ public/src/handlers/game-over.handler.js
@@ -1,30 +1,38 @@
 import packetTypes from '../constants/packet-types.constants.js';
 import Game from '../game.js';
+import { displayModalMessage } from '../modals/modal.js';
 import RequestPacket from '../protobuf/classes/request.proto.js';
 import Socket from '../socket.js';
 
+const winSound = new Audio('sounds/win.wav');
+const loseSound = new Audio('sounds/lose.wav');
+winSound.volume = 0.1;
+loseSound.volume = 0.1;
+
 const gameOverHandler = ({ socket, packetType, payload }) => {
-  // 게임 정지
+  const { isWin, baseHp } = payload;
   const game = Game.getInstance();
+
+  // 파괴 기지 최종 체력 갱신
+  if (isWin) {
+    game.opponentBase.hp = baseHp;
+    game.opponentBase.draw(game.opponentCtx, game.baseImage);
+  } else {
+    game.base.hp = baseHp;
+    game.base.draw(game.ctx, game.baseImage);
+  }
+
+  // 게임 정지
   game.end();
 
   // 결과 출력
-  const { isWin } = payload;
-  const winSound = new Audio('sounds/win.wav');
-  const loseSound = new Audio('sounds/lose.wav');
-  winSound.volume = 0.1;
-  loseSound.volume = 0.1;
-  const modal = document.getElementById('modalContainer');
-  const modalMessage = document.querySelector('.modalMessage');
   if (isWin) {
     winSound.play().then(() => {
-      modalMessage.innerHTML = '축하드립니다!. 게임에서 승리했습니다!';
-      modal.classList.remove('hidden');
+      displayModalMessage('축하드립니다!. 게임에서 승리했습니다!');
     });
   } else {
     loseSound.play().then(() => {
-      modalMessage.innerHTML = '아쉽습니다. 게임에서 패배했습니다..';
-      modal.classList.remove('hidden');
+      displayModalMessage('아쉽습니다. 게임에서 패배했습니다..');
     });
   }
